+++ cpp/11-expression_tree/node/binaryOperatorNode/binaryOperatorNode-implementation/divNode.cpp
@@ -1,9 +1,9 @@
 #include "divNode.hpp"
 
-int et::DivNode::getValue() const
+float et::DivNode::getValue() const
 {
-    int leftValue = m_left->getValue();
-    int rightValue = m_right->getValue();
+    float leftValue = m_left->getValue();
+    float rightValue = m_right->getValue();
     return leftValue / rightValue;
 
 }
