+++ cpp/11-expression_tree/node/binaryOperatorNode/binaryOperatorNode-implementation/subNode.hpp
@@ -13,8 +13,9 @@ public:
     SubNode& operator = (SubNode const& other) = delete;
     ~SubNode() = default;
 
-    int getValue() const override;
+    float getValue() const override;
 };
 
 }// namespace et
+
 #endif
